Fix queryString handling to prevent infinite loops
Before, we were always parsing the query string before passing it in as a prop.
As a result, the component would ALWAYS receive new props, even if the state
hadn't changed. Now, we only pass new props if the querystring changes, (and
then we parse it as necessary). This also allows us to use componentDidUpdate
to check the queryString