Onboarding 2.0: About page
Use an about page rather than a modal. This is slightly more flexible and
allows us to use more images to explain what is going on. Lots of big call to
action buttons (Get Started). This also fixes some of the minor cookie bugs
that we encountered. The about page still needs to be linked in a footer.